workflow/requirements.txt
langgraph>=0.0.55
langchain>=0.1.0
langchain-anthropic>=0.1.0
anthropic>=0.18.0
PyGithub>=1.59.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
typing-extensions>=4.8.0
